server updates

+++ resources/views/proponents/proponent_util.blade.php
@@ -69,8 +69,11 @@
                                 <td>{{ $row->lname .', '.$row->fname.' '.$row->mname }}</td>
                                 <td>
                                     @if($row->expired != 1)
-                                        {{ $row->actual_amount ? number_format(str_replace(',','',$row->actual_amount), 2,'.',',') :
-                                        number_format(str_replace(',','',$row->guaranteed_amount), 2,'.',',') }}
+                                        {{ $row->actual_amount 
+                                            ? number_format((float)str_replace(',', '', $row->actual_amount), 2, '.', ',') 
+                                            : number_format((float)str_replace(',', '', $row->guaranteed_amount), 2, '.', ',') 
+                                        }}
+
                                     @else
                                         0
                                     @endif
